LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/ineligible-containment-expected.txt

Test1

Test1


PASS Container ineligible for containment
FAIL Changing containment eligibility invalidates style assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 128, 0)"
